requirements.txt
annotated-types==0.7.0
blinker==1.9.0
blis==0.7.11
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.11
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
MarkupSafe==3.0.2
murmurhash==1.0.12
numpy==1.26.4
packaging==24.2
pandas==2.2.3
preshed==3.0.9
pydantic==2.11.3
pydantic_core==2.33.1
PyMuPDF==1.25.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
regex==2024.11.6
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
smart-open==6.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
thinc==8.2.5
threadpoolctl==3.6.0
tqdm==4.67.1
typer==0.9.4
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wasabi==1.1.3
weasel==0.3.4
Werkzeug==3.1.3
